+++ packages/lan/__tests__/apiv1/Patient.search.test.js
@@ -108,8 +108,8 @@ describe('Patient search', () => {
     app = await baseApp.asRole('practitioner');
 
     villages = await models.ReferenceData.findAll({ where: { type: 'village' } });
-    locations = await models.ReferenceData.findAll({ where: { type: 'location' } });
-    departments = await models.ReferenceData.findAll({ where: { type: 'department' } });
+    locations = await models.Location.findAll();
+    departments = await models.Department.findAll();
 
     await Promise.all(
       searchTestPatients.map(async ({ encounters: encountersData, ...data }, i) => {