+++ packages/lan/app/routes/apiv1/index.js
@@ -37,6 +37,7 @@ import { changePassword } from './changePassword';
 import { admin } from './admin';
 import { setting } from './setting';
 import { asset } from './asset';
+import { department } from './department';
 import { location } from './location';
 import { attachment } from './attachment';
 import { scheduledVaccine } from './scheduledVaccine';
@@ -90,6 +91,7 @@ apiv1.use('/setting', setting);
 
 apiv1.use('/asset', asset);
 
+apiv1.use('/department', department);
 apiv1.use('/location', location);
 
 apiv1.use('/attachment', attachment);