+++ packages/shared-src/src/demoData/departments.js
@@ -1,4 +1,4 @@
-import { splitIds } from './splitIds';
+import { randomRecordId, splitIds } from './utilities';
 
 export const DEPARTMENTS = splitIds(`
   Medical
@@ -15,3 +15,9 @@ export const DEPARTMENTS = splitIds(`
   Radiology
   Pharmacy
 `);
+
+export const seedDepartments = async models => {
+  const facilityId = await randomRecordId(models, 'Facility');
+  const departments = DEPARTMENTS.map(d => ({ ...d, code: d.name, facilityId }));
+  return models.Department.bulkCreate(departments);
+};