+++ packages/shared-src/src/models/index.js
@@ -43,6 +43,8 @@ export * from './PatientCarePlan';
 export * from './Setting';
 export * from './PatientCommunication';
 
+export * from './Facility';
+export * from './Department';
 export * from './Location';
 export * from './UserFacility';
 export * from './Attachment';