+++ packages/shared-src/src/models/sync/order.js
@@ -1,6 +1,9 @@
 export const MODEL_DEPENDENCY_ORDER = [
   'ReferenceData',
   'User',
+  'Facility',
+  'Department',
+  'Location',
 
   'Patient',
   'Encounter',