+++ CHANGELOG.md
@@ -4,6 +4,17 @@
 
 ### Manual steps required after upgrade ⚙
 
+- Add appropriate codes for default department and location for surveys to `packages/lan-server/local.json`:
+
+```
+"survey": {
+  "defaultCodes": {
+    "department": "xxxxxx",
+    "location": "xxxxxx",
+  },
+},
+```
+
 - Add to samoa `packages/sync-server/local.json`:
 
 ```
@@ -26,10 +37,11 @@
 
 ### Bug fixes 🐛
 
-- (left blank)
+- Fix OutpatientDischarger performance and schedule not working.
 
 ### Infrastructure and maintenance 🛠
 
+- Separate Facilities, Departments and Locations into distinct models
 - Mitigate codeship awscli tools EB version name collision, in our deployment steps
 - Deploy production meta-server from dev branch
 
@@ -93,6 +105,7 @@
 - Add PNG assistive technology device line list report
 - Add Generic Vaccine line list report
 - Allow disabling reports through server config
+- Allow running reports in child process
 
 ### Tweaks ⚖️